Use the last version of upfluence/sensu-go in the vendor
