+++ Godeps/_workspace/src/github.com/upfluence/sensu-go/sensu/check/check_request.go
@@ -1,0 +1,6 @@
+package check
+
+type CheckRequest struct {
+	*Check
+	Issued int64 `json:"issued,omitempty"`
+}
