+++ Godeps/_workspace/src/github.com/upfluence/sensu-go/sensu/client/client.go
@@ -1,0 +1,7 @@
+package client
+
+type Client struct {
+	Name          string   `json:"name"`
+	Address       string   `json:"address"`
+	Subscriptions []string `json:"subscriptions"`
+}
